travel_planner/requirements.txt
pymongo
flask
flask-cors
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
